# contracts/openapi/spec-val-001.yaml
openapi: 3.1.0
info:
  title: SPEC-VAL-001 — Validation Master Plan & Computer Software Assurance Strategy
  version: 0.1.0
  description: >
    Document 113 contract for WP-12 Document 79 (`services/gxp-api/app/modules/validation/
    commands_plan.py`, `models.py`). 3 owned entities: `validation_master_plan`,
    `validation_deliverable_requirement`, `validation_release_gate` — all PostgreSQL (GxP Core,
    authoritative).


    Document 106 row 144: `release` requires a `Released` signature from an independent QA Releaser,
    reason mandatory. `create` carries no Document 106 row — RBAC-gated only. The release gate is
    evaluated and recorded inside `release` itself (Document 79's own API list gives this entity only a
    GET); the GET below recomputes it live.
...
paths:
...
  /validation/v1/releases/{plan_id}/gate:
    get:
      operationId: getValidationReleaseGate
      tags: [validation]
      summary: Live release-gate evaluation for a plan (VAL-FR-020/024)
      x-requirement-ids: [VAL-FR-020, VAL-FR-024]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: plan_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Gate evaluation., content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseGateView' } } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'spec-gxp-001.yaml#/components/responses/NotFound' }
...
  /validation/v1/pq/{scenario_id}/signature-challenges:
    parameters:
      - { name: scenario_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: postValidationPqSignatureChallenges
      tags: [validation]
      summary: Signature challenge for a WP-12/WP-14 validation-platform signed action
      description: >
        Binds the challenge to the record id + current version + hash (SIG-FR-012/013/014). `action` selects the Document 106 meaning.
        The signature POLICY itself is unresolved until Head of Quality + Regulatory Affairs supply
        Document 106 rows (SG-172) — until then the follow-up signed command returns
        `SIGNATURE_POLICY_UNRESOLVED`, which is the correct evidenced behaviour.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010, SIG-FR-012]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignatureChallengeRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/SignatureChallengeResponse' } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404':
          description: NOT_FOUND — the referenced record does not exist.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
        '422':
          description: VALIDATION_FAILED — `action` is not an allowed action for this record.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
  /validation/v1/migrations/{run_id}/signature-challenges:
    parameters:
      - { name: run_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: postValidationMigrationsSignatureChallenges
      tags: [validation]
      summary: Signature challenge for a WP-12/WP-14 validation-platform signed action
      description: >
        Binds the challenge to the record id + current version + hash (SIG-FR-012/013/014). `action` selects the Document 106 meaning.
        The signature POLICY itself is unresolved until Head of Quality + Regulatory Affairs supply
        Document 106 rows (SG-172) — until then the follow-up signed command returns
        `SIGNATURE_POLICY_UNRESOLVED`, which is the correct evidenced behaviour.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010, SIG-FR-012]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignatureChallengeRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/SignatureChallengeResponse' } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404':
          description: NOT_FOUND — the referenced record does not exist.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
        '422':
          description: VALIDATION_FAILED — `action` is not an allowed action for this record.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
  /validation/v1/summary-reports/{report_id}/signature-challenges:
    parameters:
      - { name: report_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: postValidationSummaryReportsSignatureChallenges
      tags: [validation]
      summary: Signature challenge for a WP-12/WP-14 validation-platform signed action
      description: >
        Binds the challenge to the record id + current version + hash (SIG-FR-012/013/014). `action` selects the Document 106 meaning.
        The signature POLICY itself is unresolved until Head of Quality + Regulatory Affairs supply
        Document 106 rows (SG-172) — until then the follow-up signed command returns
        `SIGNATURE_POLICY_UNRESOLVED`, which is the correct evidenced behaviour.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010, SIG-FR-012]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignatureChallengeRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/SignatureChallengeResponse' } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404':
          description: NOT_FOUND — the referenced record does not exist.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
        '422':
          description: VALIDATION_FAILED — `action` is not an allowed action for this record.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
  /validation/v1/releases/{authorization_id}/deployment-check/signature-challenges:
    parameters:
      - { name: authorization_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: postValidationReleasesDeploymentCheckSignatureChallenges
      tags: [validation]
      summary: Signature challenge for a WP-12/WP-14 validation-platform signed action
      description: >
        Binds the challenge to the record id + current version + hash (SIG-FR-012/013/014). `action` selects the Document 106 meaning.
        The signature POLICY itself is unresolved until Head of Quality + Regulatory Affairs supply
        Document 106 rows (SG-172) — until then the follow-up signed command returns
        `SIGNATURE_POLICY_UNRESOLVED`, which is the correct evidenced behaviour.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010, SIG-FR-012]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignatureChallengeRequest' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/SignatureChallengeResponse' } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404':
          description: NOT_FOUND — the referenced record does not exist.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
        '422':
          description: VALIDATION_FAILED — `action` is not an allowed action for this record.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
...
  /validation/v1/releases/{vsr_id}/authorize/signature-challenges:
    parameters:
      - { name: vsr_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      operationId: postValidationReleasesAuthorizeSignatureChallenges
      tags: [validation]
      summary: Signature challenge for a WP-12/WP-14 validation-platform signed action
      description: >
        Content-hash-bound signed-CREATE (Document 106 row 166). The body must carry the exact field values later submitted to the `authorize` command — any difference invalidates the challenge.
        The signature POLICY itself is unresolved until Head of Quality + Regulatory Affairs supply
        Document 106 rows (SG-172) — until then the follow-up signed command returns
        `SIGNATURE_POLICY_UNRESOLVED`, which is the correct evidenced behaviour.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010, SIG-FR-012]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueReleaseAuthorizationChallengeBody' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/SignatureChallengeResponse' } } }
        '401': { $ref: 'spec-gxp-001.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'spec-gxp-001.yaml#/components/responses/Forbidden' }
        '404':
          description: NOT_FOUND — the referenced record does not exist.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
        '422':
          description: VALIDATION_FAILED — `action` is not an allowed action for this record.
          content: { application/json: { schema: { $ref: 'spec-gxp-001.yaml#/components/schemas/ErrorResponse' } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ReleaseGateView:
      type: object
      additionalProperties: false
      required: [plan_id, plan_state, blockers, gate_state, required_artifacts]
      x-requirement-ids: [VAL-FR-020, VAL-FR-024]
      properties:
        plan_id: { type: string, format: uuid }
        plan_state: { type: string }
        blockers: { type: array, items: { type: string } }
        gate_state: { type: string, enum: [PASSED, BLOCKED] }
        required_artifacts: { type: array, items: { type: string } }
...
    SignatureChallengeRequest:
      type: object
      additionalProperties: false
      description: Body of a validation-platform signature-challenge endpoint.
      x-requirement-ids: [SIG-FR-005]
      required: [action]
      properties:
        action: { type: string, description: One of the record's allowed signed actions (e.g. release / approve / create / deployment_check). }
    IssueReleaseAuthorizationChallengeBody:
      type: object
      description: >
        The exact field set later submitted to `POST /releases/{vsr_id}/authorize`; the challenge hash
        is computed over it, so any later difference invalidates the challenge. Shape mirrors
        `IssueValidatedReleaseAuthorizationCommand` (minus `idempotency_key`); not closed here because
        that command's field set evolves with the validation-platform build.
      x-requirement-ids: [SIG-FR-005, SIG-FR-010]
    SignatureChallengeResponse:
      type: object
      description: The challenge receipt (response schema, not closed).
      x-requirement-ids: [SIG-FR-005, SIG-FR-027]
      required: [challenge_id, meaning, expires_at]
      properties:
        challenge_id: { type: string, format: uuid }
        meaning: { type: string }
        expires_at: { type: string, format: date-time }
...